requirements.txt
numpy==1.26.4
pandas==1.5.3
scikit-learn==1.5.0
xgboost==2.1.1
torch==2.4.1
lightning==2.1.4
optuna==4.6.0
optuna-integration==4.7.0
pyyaml